fix(fe): move course notice page under (overview)/notice route
